Connect to GtkWidget::hide instead of various GtkMenuShell signals
In some contexts GtkMenuShell::selection-done is emitted while
GtkMenuShell::deactivate is not, or vice versa, especially on Wayland
(but also on X11). Furthermore, all signal handlers connected to either
of these signals actually contain actions to be performed when the menu
is hidden, regardless of why it is hidden (it can also be destroyed in
fact, as provided in `panel-plugin.c`).

So in summary the right signal to connect to, both for simplicity and to
always get the desired result, is GtkWidget::hide.