Connect to GtkMenuShell::deactivate instead of selection-done
Except for the application menu, where it is necessary to connect to
GtkMenuShell::selection-done which is the only signal emitted when a dnd
is cancelled, GtkMenuShell::deactivate appears as a generic signal to
perform operations when closing the menu. This is already the one that
is used most of the time in the code.

Moreover, on Wayland, GtkMenuShell::deactivate is the only signal
emitted when clicking outside the menu, which is a very frequent
situation. So it seems desirable to connect only to this signal.

However, it is then necessary to postpone the destruction of the menu if
it takes place in the signal handler, as it is already the case in other
places in the code, so that a possible activation of a menu item has
time to finish.