tasklist: Do not duplicate menu item activation
Connecting to "button-press-event" is just useless, but connecting to
"button-release-event" should be kept for middle click actions. Luckily
the left click worked on X11 because the two activations seem to work
together, but they cancel each other out on Wayland. As for the right
click, it did not work either on X11 or on Wayland.