clock: Do not override custom font formats
The font specified via the font dialog can be applied globally while
giving priority to custom font formats, as long as it is applied at the
end and without mixing markup strings and Pango attributes, as the doc
recommends. This means not applying the dialog font and the custom
format separately, but all at once, so some refactoring.

Fixes: #675
Related: !101