pager: Use WnckHandle if available
In the current state of the code on the master branch, this is probably
not necessary, as Libxfce4windowing >= 4.19.1 no longer uses wnck icons.
But it does reduce the distance from the 4.18 branch, where this change
is made necessary by the use of `wnck_set_default_icon_size()` in
tasklist and windowmenu.

Also, if plugins are forced to run as internal via "force-all-internal",
it is possible that one of them will change this icon size. And in any
case, it is probably better to use this API when possible, since it
allows to fix a deprecation.

Related: #695