Revert "Use RGBA visual only if compositing is enabled (Fixes #251)"
This reverts commit 87c609e7a576425f69b4c56786dbf657c278ae6a.

While this fixed #251, it had the side effect of forcing a restart of
the panel to get the transparency working if compositing was not enabled
before (since gtk_widget_set_visual() must be called before the panel is
realized). Since #251 is a rather special case, the lesser evil seems to
be to use the RGBA visual as soon as possible.

Fixes: #672
(cherry picked from commit 24af49d79c1b7ec92a41d524c22da25ed4f216ae)