cleanup: Revert use of G_DECLARE_FINAL_TYPE for PanelBaseWindow
It worked as is but it's misleading since PanelBaseWindow is derivable.
A clean refactoring would be to use G_DECLARE_DERIVABLE_TYPE by
replacing the public fields of the instance structure with accessors and
`priv` with `get_instance_private()`. But that will be for later, as for
PanelPluginExternal in a previous commit.

Fixes: c7f32ce82f01738948dc522c2a48881d05c33341