panel: Delay removal of ExternalPlugin to prevent use-after-free
The sanity check in `panel_plugin_external_child_watch_destroyed()` was
probably intended to address this problem, but it is inappropriate.
While it serves its purpose in many configurations by failing and thus
avoiding a pointer dereference that is more likely to cause a crash, it
can itself cause a crash during type checking and ultimately makes the
problem harder to debug.

The right thing to do is to make sure that you don't get into the idle
functions once finalized, here by extracting the PluginExternal removal
from these functions (as the handling of the source id seems otherwise
correct).

Fixes: #711