libxfce4panel: Fix memory management of source for menu positioning
This is unfortunately not safe as it is, because if two pointer events
to show and hide the menu are close enough, the second one will be
processed before the idle func is executed. And if hiding the menu
triggers its destruction, it causes a use-after-free. This is for
example the case with the DirectoryMenu plugin.

Fixes: 03f01527a3c7bd8c03ad79788476a99d36b35348