panel: Do not reset output name if a monitor does not have a model name
There seems to be no good reason to do this. On the one hand, the fact
that a monitor does not have a model name does not prevent another
monitor from having one, whether it is currently connected or not. In
particular, it is not clear that the fact that `gdk_monitor_get_model()`
returns NULL means that "the driver does not support output names" as
the error message says. Also, it seems that there may be situations
where this request occurs too early, so that `gdk_monitor_get_model()`
returns NULL only temporarily.

On the other hand, contrary to what the error message says, this reset
of output name has no effect on the saved setting, since the
"output-name" property is not informed about it. So that once passed in
this code path, the internal state of the panel does not even correspond
to what the user can observe in the panel preferences. Anyway, the
settings entered by the user should not be changed without his consent
in general.

Fixes: https://gitlab.xfce.org/xfce/xfce4-panel/-/issues/151#note_66464