libxfce4panel: Unregister menu also on GtkWidget::hide
There may be cases where the menu is hidden manually, and thus without
emitting "selection-done" or "deactivate". Example: the launcher menu
when dnding from Thunar.

It seems more unlikely that "selection-done" or "deactivate" is emitted
without "hide" being emitted, however a manual emission of one of these
signals on a menu that would never have been shown can't be totally
excluded, so it's better to keep the existing connections rather than
replace them. The same applies to "destroy".

Fixes: #716