wayland: Commit wl_surface manually only if necessary
Second side effect of previous commit, which is combined with a side
effect of manual wl_surface commit: the latter causes the pointer to
re-enter the autohide window when it is moved off-screen, which can lead
to an autohide cycle that causes the panel to flicker due its remap.
Easier to reproduce without external plugins.

Related: 4bb886d0ae54ef424887aad49d4b8abf54caca00