panel: Rephrase "Don't reserve space on borders" (V2)
First reworded as "Keep panel above windows" in 6f64ea01 but it
obviously doesn't meet with consensus, so new rewording as "Reserve
space on screen edges for the panel", more in line with the initial
request and more technically correct.

This time it requires a setting migration for backwards compatibility
though, since the toggle is reversed.

Closes: #217