launcher: Show action menu also when there are several items
Maybe it's a bit strange to see the actions of the first element when
you click on the launcher arrow, but it's nice that they appear when you
click on the element itself. So let's keep it simple and show them all
the time.

Fixes: #720
Related: 7f3dd44de951a74dc1adb446e2b2f643043b1100