clock: Fix sync timeout for seconds display
5cf7006e fixed the same problem for the minutes display, but there's a
bit more work here, since nothing was done in addition to
`g_timeout_add()` to synchronize the seconds display.

A synchronization every 10 seconds with a precision of a tenth of a
second seems reasonable: this gives a readjustment every 50 to 60
seconds on my system, if the panel isn't too busy (since the plugin is
internal). I took the opportunity to raise the timeout priorities, which
should improve things a little in this case.

Fixes: #756
Completes: 5cf7006ee02cac06a693714ee9c30d3e7d326c17