panel: Remove autosave timer for plugins
It was introduced in 89cfd529 without rationale, and it's not clear that
a save signal should be sent to plugins every ten minutes. On the
contrary, it seems that this should only be done on rare occasions,
typically on shutdown, which is retained. For the rest, it's up to each
plugin to manage its backups as it sees fit.

Fixes: #767