pager: Do not adapt upper value of spin button in viewport mode
This doesn't make sense. Libwnck handles this internally without any
obvious way for us to get the number of rows, so it's best not to allow
the user to change the number of rows in this case, as this will only
lead to bad results.

(cherry picked from commit 390d152aae6fe3d5b1cad813e3a39ce9799dbf60)