actions: Populate plugin earlier
This is necessary on Wayland with the current external plugin mechanism
for some reason, otherwise the plugin is never allocated (most of the
time). This is reminiscent of cases where the panel must be remapped to
trigger allocation (see panel_utils_widget_remap() use cases), besides
remapping also fixes the problem in this case.

Anyway, raising the idle func priority works in this case while
retaining the benefit of delayed execution, so let's do that.