Try to derive plugin dirs from XDG_DATA_DIRS
The aim is not to cover all cases, or even as many as possible, but
simply to make life easier for plugin developers and/or users in these
fairly typical situations:
* Installing a plugin in /usr/local when the panel is installed in /usr
  -> currently, the plugin doesn't appear in the add items dialog.
* Installing an alternative version of the panel in /usr/local while the
  system version and its plugins are installed in /usr -> currently, the
  alternative version of the panel loses all its plugins on startup, as
  it cannot find them in /usr/local (#718).

If the panel's datadir and libdir installation directories are
conventional, i.e. of the form $prefix$data_suffix and
$prefix$lib_suffix, then this commit should cover the above cases. More
generally and precisely, it searches among the XDG_DATA_DIRS and
XDG_DATA_HOME directories for those ending in $data_suffix, and tries to
use their $_prefix to find plugins in the pair
($_prefix$data_suffix, $_prefix$lib_suffix).

This also drops support for the old plugin location, which was
deprecated in 2010.

Related: https://bugzilla.xfce.org/show_bug.cgi?id=5455
Helps: #718