tasklist: Sanitize XfwApplication management
Now that we're using our own library, we can ensure that XfwApplication
is non-null and stays alive as long as one of its windows is alive, so
we don't need to take a reference from it. What's more, the app table
doesn't need to exist if grouping by application is not used. So we no
longer intervene in the lifecycle of XfwApplication, which is released
in libxfce4windowing as soon as its last window is closed.