tasklist: Fix default button length
Since 0f43d3f8, the length of the buttons varies between
min-button-length and max-button-length depending on the length of the
labels, which is fine and complies with the documentation, but may not
be what all users want. Setting a minimum length equal to the maximum
length ensures the same default behavior as before, overflow menu aside.

Amends: 0f43d3f8e028b4af9220abb4a840d8ee7ac278fe