prefs-dialog: Improve launcher display in item list
If this were to be done for many others, or even a few other plugins, it
would be worth adding an API allowing plugins to provide better names
and icons than generic ones, but it seems really launcher-specific. On
the other hand, it's a plugin shipped with the panel for which this
enhancement seems really relevant, so it seems to justify this little
bit of custom code.

Closes: #21