wayland: Avoid allocation loop when the panel is anchored to an edge
This hack is still necessary to be able to resize the panel correctly,
but it has triggered an allocation loop in labwc since version 0.8.0 (in
fact since the use of wlroots 0.18 just after 0.7.3). Since it's only
needed when the panel changes size, it's better to move it there and
restrict its use to that case anyway.