autohide: Show/hide autohide window instead of moving it offscreen
Moving layer-shell surfaces off-screen on Wayland is problematic, as it
requires remapping, which causes unwanted blinking and pointer events,
and manual commit, which can cause protocol errors. So let's try not to
do that as much as possible, while preserving common code paths between
X11 and Wayland.

First step: autohide window (to catch pointer events when the panel is
hidden). As there is no animation on this window, we can simply hide it
on X11 as on Wayland.