wayland: autohide: Show/hide panel instead of moving it offscreen
Second step: we hide the panel window on Wayland when it's about to be
moved off-screen, either directly or at the end of the animation (the
same applies to external plugins). This eliminates the need for
remapping and manual commit, since as long as the window has not
completely left the screen, it can be brought back without remapping,
and remapping is not necessary either to trigger a resize. See
https://github.com/labwc/labwc/issues/499

It might be preferable to manage animation on Wayland via the viewporter
protocol, as the layer-shell protocol may no longer support negative
margins when it arrives in the ext namespace. Unfortunately, the current
implementation of the viewporter protocol in most compositors doesn't
allow this: see https://github.com/labwc/labwc/issues/2129