wayland: autohide: Do not block when panel has input focus
* The panel gains focus as soon as it is shown, or on various other
  occasions.
* It depends on the compositor.
* It depends on whether or not there are external plugins (in fact,
  other layer-shell surfaces competing for focus).

For all these reasons and perhaps others, it's not something you can
rely on to block autohide. In practice, it could at least require the
user to manually give focus to another window so that the panel hides
(instead of hiding itself when the pointer leaves it), or even trigger
focus loss/gain loops where the panel fails to hide itself.