wayland: Avoid allocation loop when the panel is anchored to an edge
Second try after b7083b35: reset anchors to default only if the natural
width/heght change, so this should break the allocation loop and
preserve auto-adjustement.

Note that it's also necessary to delay the anchor change rather than
apply it between get_preferred_(width|height) and size_allocate, as this
can cause protocol errors (the internal workings of allocation depend on
the anchor type, so we shouldn't change it on the fly). On the other
hand, you have to re-apply anchors as soon as possible, otherwise
maximized windows may flicker (exclusive zone is disabled when the panel
is only anchored to two edges).