wayland: tasklist: Fix monitor detection when panel is remapped
gdk_display_get_monitor_at_window() is unreliable on wayland, because
when the panel is remapped it usually first appears on one monitor
before being moved to another, and there is no signal or event to
indicate when this process has finished. So it's best to use
gtk_layer_get_monitor() whenever possible, as we always use
gtk_layer_set_monitor() upstream.