LCD and digital clock layout fixes
Maked size/layout of the lcd clock time-independent

LCD clock had a feature where the size of the plugin
could be reduced depending on time of the day.
As a side-effect, other plugins were shifted when the size
of the LCD clock changes.
This commit removes this feature, fixing the clock size.

Made digital clock lines not to spread out in wide panels.