Reworked sizing mechanism in clock plugin and its widgets
Original mechanism relied on a "ratio" property of clock widgets
and the clock plugin setting a fixed size request. This mechanism
was fragile and resulted in missing size updates. Changed this mechanism
to width_for_height/height_for_width layout. Removed no longer needed
ratio property and added a container-orientation one (the widget has to
be aware of the panel/container orientation for layout purposes).

Also combined the change with size clamping of the LCD and Binary widgets.
Not ideal but it was strongly coupled with the layout mechanism.