item-dialog: Also search for plugins by their module name
It's not uncommon to know a plugin by the (untranslated) name under
which it was installed, rather than by its display name, or the words in
its description (both translated). The module name is essentially
(sometimes exactly) the name under which the plugin was installed, which
should make it easier to find.