itembar: Fix memory leak mitigation on plugins
A reference widget must be allocated for each plugin instead of using a
static variable, otherwise when several plugins are removed at once
(e.g. when removing a panel), the memory leak mitigation only applies to
the first of them.

On the other hand, panel destruction must also be deferred, so that
plugins can refer to it when they are destroyed.

Fixes: 6b671b0d03a353fae93ff09a769cab27a66018ac