wayland: autohide: Try to trigger size_allocate() more reliably
It seems to me that it worked like that "before", i.e. probably with a
previous version of wlroots. I guess this isn't the last time I'll have
to tinker with this, but I'd really like to try not to remap the panel
anyway.

Amends: ee22684ba748706c39440d897c78e0fcde813fb4